Handle Google Sing up issues and error handling
