+++ client/src/app/components/App/Shared/GoogleButton/googleLoginButton.scss
@@ -1,4 +1,4 @@
-.google-login-button {
+.google-login-button, .google-signup-button {
   display: block;
   border: none;
   border-radius: 0;
